+++ Qtica/widgets/window/__init__.py
@@ -1,2 +1,3 @@
+from .base import BaseWindow
 from .mainwindow import MainWindow
 from .frameless import FramelessWindow